fix(dsacolor): generate wider range of colors
color type and std::rand() were not big enough to generate red component

+++ include/seadsa/DsaColor.hh
@@ -10,7 +10,7 @@ using namespace llvm;
 
 namespace seadsa {
 
-using Color = unsigned;
+using Color = unsigned long long; // at least 64 bit, needed 48
 
 class Graph;
 class Node;
